Move the test interface inside of `src`
